backend/requirements.txt
psycopg2-binary==2.9.3
python-dateutil==2.8.2
Django==4.1.1
pytest-django==4.5.2
django-ninja==0.19.1